requirements.txt
-e .
PyFoam
black
isort
openpyxl
matplotlib
seaborn
pandas
pytest
pyyaml
pylint
psycopg2-binary
umap-learn
scikit-learn
sagemaker
optuna